KMarkovGenSeq

See our report in Report

See the performance and benchmark test in Tests

APIs

KMarkovGenSeq.db_freq — Method

Process the counts of each k-strr in several database sequences. And the -log frequency will be given for the origin sequences and their reverse complementary sequences.

The first value to return is the descriptions of every database sequences. The second is the -log frequencies.

Args:

dbseqs: the forlder of the database sequences

len: do K-Markov algo with K = len

KMarkovGenSeq.query_freq — Function

Perform K-Markov algorithm to predict where sequences come from.

Args:

query_seqs: the file containing the query sequences

dbseqs: the forlder of the database sequences

len: do K-Markov algo with K = len

out_file: the output file to give the predictions, default is nothing and no file will be given

use_gpu: whether to use gpu